+++ src/languages/capnproto.js
@@ -8,6 +8,7 @@ Category: protocols
 
 export default function(hljs) {
   return {
+    name: 'Cap’n Proto',
     aliases: ['capnp'],
     keywords: {
       keyword: