+++ src/languages/clojure.js
@@ -97,6 +97,7 @@ export default function(hljs) {
   HINT_COL.contains = [COLLECTION];
 
   return {
+    name: 'Clojure',
     aliases: ['clj'],
     illegal: /\S/,
     contains: [LIST, STRING, HINT, HINT_COL, COMMENT, KEY, COLLECTION, NUMBER, LITERAL]