+++ src/languages/elm.js
@@ -46,6 +46,7 @@ export default function(hljs) {
   };
 
   return {
+    name: 'Elm',
     keywords:
       'let in if then else case of where module import exposing ' +
       'type alias as infix infixl infixr port effect command subscription',