+++ src/languages/sqf.js
@@ -64,6 +64,7 @@ export default function(hljs) {
   };
 
   return {
+    name: 'SQF',
     aliases: ['sqf'],
     case_insensitive: true,
     keywords: {