+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 New languages:
 
-- none.
+- enh(python) Added `python-repl` for Python REPL sessions
 
 New themes:
 
@@ -17,6 +17,7 @@ Core Changes:
 Language Improvements:
 
 - enh(fortran) Support Fortran 77 style comments (#2416) [Josh Goebel][]
+- (csharp) add support for `@identifier` style identifiers (#2414) [Josh Goebel][]
 - fix(elixir) Support function names with a slash (#2406) [Josh Goebel][]
 - fix(javascript) comma is allowed in a "value container" (#2403) [Josh Goebel][]
 - enh(apache) add `deny` and `allow` keywords [Josh Goebel][]